Windows: Refactor volumes
