+++ daemon/volumes_windows.go
@@ -1,0 +1,8 @@
+// +build windows
+
+package daemon
+
+// Not supported on Windows
+func copyOwnership(source, destination string) error {
+	return nil
+}
